Move logic to the lib folder
